Compiled out process.env.NODE_ENV for Rax and React benchmarks
